config now handles multiple buttons
